Fix merging of validations
Normally, Rails's validate method clears existing validations before
running. However, as we are setting some errors based on the wider
context at the top level, this resulted in errors being clobbered when
we recursed down.

Luckily, because we never use validate directly, but through a
controlled process, we don't need to clear them.